// src/model/Response/Hotel/HotelGetAllRes.tsx
export interface HotelGetAllRes {
    HID: number;
    hotel_user_ID: number;
    name_user: string;
    hotel_type_ID: number;
    typename_hotel: string;
    name: string;
    province: string;
    address: string;
    detail: string;
    latitude: number;
    longtitude: number;
    datetime_addhotel: Date;
  }